refactoring & style changes
