+++ webapp/app/results/[generation_id]/layout.tsx
@@ -1,8 +1,3 @@
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
-
 export default function RootLayout({
   children,
 }: {
