+++ webapp/app/results/layout.tsx
@@ -1,6 +1,6 @@
 export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
+  title: 'Results',
+  description: "Stable Diffusion and LRP",
 }
 
 export const dynamic = 'force-dynamic'
