+++ webapp/app/results/page.tsx
@@ -1,4 +1,4 @@
-import FetchGenerations from '@/components/generation/list_generations_server/ListGenerationsServer';
+import FetchGenerations from '@/components/generation/list_generations/ListGenerationsServer';
 import UploadAssets from '@/components/upload_assets/UploadAssets';
 
 export const dynamic = 'force-dynamic'
