+++ webapp/app/unet/layout.tsx
@@ -1,6 +1,6 @@
 export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
+  title: 'Unet',
+  description: "Stable Diffusion and LRP",
 }
 
 export default function RootLayout({
